feat: dynamiclist with filter
