+++ src/DynamicTreeDataGrid/DynamicFlatTreeDataGridSource.cs
@@ -43,11 +43,13 @@ public class DynamicFlatTreeDataGridSource<TModel, TModelKey> : FlatTreeDataGrid
         // TODO: Setup Sorted event for treeDataGridSourceImplementation?
     }
 
+
+
     public IObservable<int> FilteredCount { get; }
     public IObservable<int> TotalCount { get; }
 
     public new DynamicColumnList<TModel> Columns { get; }
-    IColumns ITreeDataGridSource.Columns => Columns;
+    IColumns ITreeDataGridSource.Columns => Columns.DisplayedColumns;
 
     // TODO: Change to check the sort observable.
     // public bool IsSorted => _comparer is not null;
