+++ src/DynamicTreeDataGrid/Models/Columns/IDynamicColumn.cs
@@ -12,5 +12,5 @@ public interface IDynamicColumn : IColumn {
 	// /// <summary>
 	// /// Whether the column should be visible in the table. Defaults to <c>true</c>
 	// /// </summary>
-	// public bool Visible { get; set; }
+	public bool Visible { get; set; }
 }
